services/catalyst/pyproject.toml
[project]
name = "catalyst"
version = "0.1.0"
description = "Catalyst — Product Launch Intelligence Spoke for OctaneBrew Platform"
readme = "README.md"
authors = [
    { name = "shubh305", email = "shubham.305singh@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "pydantic-settings>=2.12.0",
    "httpx>=0.28.1",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "elasticsearch==9.2.1",
    "aiohttp>=3.9.0",
    "redis[hiredis]>=5.0.1",
    "aiokafka>=0.13.0",
    "confluent-kafka>=2.3.0",
    "pandas>=2.0.0",
    "kaggle>=1.6.0",
    "praw>=7.7.0",
    "tenacity>=8.2.3",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-instrumentation-httpx>=0.43b0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
catalyst = "catalyst:main"

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"
